requirements.txt
# ============================================================
# PolicyBot v4 — Python Requirements
# ============================================================
# Install command:
#   pip install -r requirements.txt
#
# Also required (system package, NOT pip):
#   Windows : https://github.com/UB-Mannheim/tesseract/wiki
#             Install path: C:\Program Files\Tesseract-OCR\tesseract.exe
#   Ubuntu  : sudo apt install tesseract-ocr poppler-utils
#   macOS   : brew install tesseract poppler
# ============================================================

# ── Web Framework ────────────────────────────────────────────
Flask==3.1.2
flask-cors==4.0.1
Werkzeug==3.1.5
Jinja2==3.1.6
itsdangerous==2.2.0
click==8.3.1

# ── Environment & Config ─────────────────────────────────────
python-dotenv==1.2.1

# ── Gemini AI (conversation + optional Vision fallback) ──────
google-generativeai==0.7.2

# ── Offline OCR — Image Processing ──────────────────────────
pytesseract==0.3.13
opencv-python==4.13.0.92
Pillow==12.1.1
numpy==2.4.2

# ── Document Reading ─────────────────────────────────────────
pypdf==5.9.0
pdf2image==1.17.0
python-docx==1.2.0
pi
# ── Production Server ────────────────────────────────────────
gunicorn==22.0.0

# ── Optional: Vector DB for RAG ──────────────────────────────
faiss-cpu==1.8.0
chromadb==0.5.0